Change PortWithProtocol to PortsWithProtocol
